pyproject.toml
[tool.poetry]
name = "bearcoderr"
version = "1.0.1"
description = "Project Alekseii Achkasov"
authors = ["Alekseii Achkasov <bearcoderr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
django-ckeditor = "^6.7.1"
reportlab = "^4.2.5"
celery = "^5.4.0"
redis = "^5.2.0"
django-celery-results = "^2.5.1"
feedgenerator = "^2.1.0"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
